+++ knowage-vue/src/modules/documentExecution/dashboard/widget/ChartWidget/classes/highcharts/KnowageHighcharts.ts
