+++ knowage-vue/src/modules/documentExecution/dashboard/widget/ChartWidget/classes/highcharts/KnowageHighchartsGaugeChart.ts
@@ -1,7 +1,7 @@
 import { IWidget, IWidgetColumn } from '@/modules/documentExecution/dashboard/Dashboard'
 import { IHighchartsGaugeSerie, IHighchartsGaugeSerieData } from '@/modules/documentExecution/dashboard/interfaces/highcharts/DashboardHighchartsGaugeWidget'
 import { IHighchartsChartSerie } from '@/modules/documentExecution/dashboard/interfaces/highcharts/DashboardHighchartsWidget'
-import { KnowageHighcharts } from './KnowageHihgcharts'
+import { KnowageHighcharts } from './KnowageHighcharts'
 import { createGaugeSerie } from './updater/KnowageHighchartsCommonUpdater'
 
 export class KnowageHighchartsGaugeChart extends KnowageHighcharts {
